Move GroupWithLights to separate package.
