+++ app/src/main/java/org/d3kad3nt/sunriseClock/data/model/groupWithLights/DbGroupLightCrossref.java
@@ -1,4 +1,4 @@
-package org.d3kad3nt.sunriseClock.data.model.group;
+package org.d3kad3nt.sunriseClock.data.model.groupWithLights;
 
 import androidx.room.ColumnInfo;
 import androidx.room.Entity;
@@ -6,6 +6,7 @@ import androidx.room.ForeignKey;
 import androidx.room.Ignore;
 import androidx.room.Index;
 
+import org.d3kad3nt.sunriseClock.data.model.group.DbGroup;
 import org.d3kad3nt.sunriseClock.data.model.light.DbLight;
 
 @Entity(tableName = DbGroupLightCrossref.TABLENAME,
